requirements.txt
bcrypt==4.2.0
cffi==1.16.0
cryptography==43.0.0
getmac==0.9.5
joblib==1.4.2
netifaces==0.11.0
ntplib==0.4.0
numpy==2.0.1
pandas==2.2.2
paramiko==3.4.0
prettytable==3.10.2
pycparser==2.22
PyNaCl==1.5.0
python-dateutil==2.9.0.post0
python-nmap==0.7.1
pytz==2024.1
scapy==2.5.0
scikit-learn==1.5.1
scipy==1.14.0
six==1.16.0
threadpoolctl==3.5.0
tzdata==2024.1
wcwidth==0.2.13
